webapp/requirements.txt
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
python-mimeparse
six
Werkzeug
mecab-python3
